impl priviledge check for entity manipulation

+++ libs/api/comments/src/lib/comments.service.ts
@@ -1,6 +1,12 @@
 import { Comment, Post, User } from '@ehc/common/entities';
 import { CreateComment, UpdateComment, CommentQuery } from '@ehc/common/dtos';
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  Logger,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
@@ -27,26 +33,40 @@ export class CommentsService {
     return comment;
   }
 
-  create(dto: CreateComment): Promise<Comment> {
+  create(dto: CreateComment, caller: User): Promise<Comment> {
     const comment = this.commentsRepository.create(dto);
-    comment.author = { id: dto.authorId } as User;
+    comment.author = { id: caller.id } as User;
     comment.post = { id: dto.postId } as Post;
     return this.commentsRepository.save(comment);
   }
 
-  async update(id: number, dto: UpdateComment): Promise<Comment> {
-    const comment = await this.commentsRepository.findOneBy({ id });
+  async update(id: number, dto: UpdateComment, caller: User): Promise<Comment> {
+    const comment = await this.commentsRepository.findOne({
+      where: { id },
+      relations: { author: true },
+    });
     if (!comment) {
       throw new NotFoundException(`Comment with id ${id} was not found`);
+    }
+    if (comment.author.id != caller.id) {
+      throw new ForbiddenException();
     }
     await this.commentsRepository.update(id, dto);
     return (await this.commentsRepository.findOneBy({ id }))!;
   }
 
-  async delete(id: number): Promise<Comment> {
-    const comment = await this.commentsRepository.findOneBy({ id });
+  async delete(id: number, caller: User): Promise<Comment> {
+    const comment = await this.commentsRepository.findOne({
+      where: { id },
+      relations: {
+        author: true,
+      },
+    });
     if (!comment) {
       throw new NotFoundException(`Comment with id ${id} was not found`);
+    }
+    if (comment.author.id != caller.id) {
+      throw new ForbiddenException();
     }
     this.commentsRepository.delete(id);
     return comment;
